+++ src/libtropic.c
@@ -1135,7 +1135,7 @@ lt_ret_t lt_ecc_key_store(lt_handle_t *h, const ecc_slot_t slot, const lt_ecc_cu
 }
 
 lt_ret_t lt_ecc_key_read(lt_handle_t *h, const ecc_slot_t ecc_slot, uint8_t *key, lt_ecc_curve_type_t *curve,
-                         ecc_key_origin_t *origin)
+                         lt_ecc_key_origin_t *origin)
 {
     if (!h || (ecc_slot > ECC_SLOT_31) || !key || !curve || !origin) {
         return LT_PARAM_ERR;